requirements.txt
opencv-python
numpy
PyExifTool
tqdm